Add ssr for supabase
