+++ src/app/components/events/EventsClient.tsx
@@ -10,7 +10,7 @@ import ColouredContentBox from "../common/ColouredContentBox"
 import { useRouter, usePathname, useSearchParams } from "next/navigation";
 import SingleEventModal from "./SingleEventModal";
 import { useEffect } from "react";
-import { EventsData } from "@/services/event.service";
+import { EventsData } from "@/services/event.client.service";
 import { FaChevronLeft} from "react-icons/fa6"
 import { FaChevronRight } from "react-icons/fa6"
